feat(taxonomy): export Siren's full typed taxonomy; wire mypy strict lint
Siren owns its domain vocabulary (global-taxonomy-and-sdk-doctrine); this SDK
is the typed carrier of it so integrations never hand-roll magic strings.

- Add EventSlug (built-in ingestion slugs) and status enums:
  ConversionStatus, TransactionStatus, ObligationStatus, PayoutStatus,
  FulfillmentStatus, OpportunityStatus, ApiKeyStatus,
  WebhookSubscriptionStatus (all str-valued, exported from the package root).
- WebhookEventType: add CREDIT_ISSUED, CREDIT_REDEEMED, CURRENCY_CREATED,
  CURRENCY_DELETED — dispatched by the service but missing from the SDK and
  openapi.yaml.
- Pin the canonical vocabularies in tests/test_taxonomy.py (mirrors the
  service's dispatchers and REST validations).
- Wire mypy --strict as the repo lint (pyproject [tool.mypy], navigator.yaml
  ci.lint, CI lint job); fix the two findings it surfaced.

Charter 761 (SDKs speak the taxonomy).

+++ src/siren/_client.py
@@ -87,7 +87,7 @@ class HttpClient:
     def _backoff_delay(self, attempt: int, error: SirenError) -> float:
         if isinstance(error, RateLimitError) and error.retry_after is not None:
             return float(error.retry_after)
-        return min(_INITIAL_BACKOFF * (2**attempt), _MAX_BACKOFF)
+        return float(min(_INITIAL_BACKOFF * (2**attempt), _MAX_BACKOFF))
 
 
 def parse_json(response: httpx.Response) -> Any:
